refactor(Server .cpps): create new ServerWrappers.cpp & ServerErrorHandling.cpp for better organization. Add comments to functions. Rename getAddrInfo() param for clarity
