+++ Makefile
@@ -12,7 +12,9 @@ SRC			+= Server.cpp\
 			ServerInit.cpp\
 			ServerRun.cpp\
 			ServerPollHandling.cpp\
-			ServerDebug.cpp
+			ServerErrorHandling.cpp\
+			ServerDebug.cpp\
+			ServerWrappers.cpp
 vpath %.cpp $(SRC_DIR)/logging
 SRC			+= Logging.cpp
 vpath %.cpp $(SRC_DIR)/config
