+++ src/server/ServerDebug.cpp
@@ -1,21 +1,21 @@
 #include "Server.hpp"
 #include <cerrno>  // for errno
 #include <cstring> // for strerror
-// #include <fcntl.h> // for demonstration purposes only, forbidden function
 
-// get_addr_info_str() is only for logging purposes.
-// The function builds a string from the contents of an addrinfo struct
-// and an additional msg, which I've used to add a label for this struct.
-//
-// NOTE: This function should not be used with uninitialized addrinfo structs,
-// in which the values might be garbage. Currently, it is only used by
-// getServerInfo(), which assures initialization.
-//
-// Currently static due to one-time use. In the future, may be useful for
-// logging client addrinfo, in which case it would need to be added back
-// to the header, and could be moved to a different (debug / helper?) .cpp
-//
-// RETURNS: a string with all info to be printed
+// #include <fcntl.h> // for demonstration purposes only
+// fcntl() is not an allowed function, but can be used to demonstrate
+// that client sockets are blockinb by default. Uncomment printFcntlFlags()
+
+/**
+* \brief Converts addrinfo struct to a string for logging purposes.
+*
+* Currently, it is only used by getServerInfo().
+*
+* \param    Info pointer to an addrinfo struct
+* \param    Msg a string to be included in the output, i.e. a label / title
+*
+* \return a string with all info to be printed
+*/
 
 std::string Server::addrinfoToStr(const struct addrinfo *Info,
                                   const std::string &Msg) {
@@ -29,33 +29,11 @@ std::string Server::addrinfoToStr(const struct addrinfo *Info,
   return (oss.str());
 }
 
-static std::string interfaceInfoToStr(const Listen &Interface) {
+std::string Server::interfaceInfoToStr(const Listen &Interface) {
   std::ostringstream msg;
       msg << "\n" << "IP: " << Interface.ip << "\n"
       << "Port: " << Interface.port << "\n\n";
   return (msg.str());
-}
-
-void Server::handleBindFailure(const struct addrinfo *Info, const Listen &Interface, const int Error) {
-  std::ostringstream msg;
-  msg << "bind: ";
-      switch (Error) {
-      case EACCES:
-        msg << "EACCESS: " << std::strerror(Error) << "\n" << interfaceInfoToStr(Interface)
-            << "Use a port number above 1024, or run webserv with root privileges.\n";
-        break;
-      case EADDRNOTAVAIL:
-        msg << "EADDRNOTAVAIL: " << std::strerror(Error) << "\n" << interfaceInfoToStr(Interface)
-        << "Check that the IP address in the config file is correct for this machine\n"
-        << "Use 127.0.0.1 for local host, or find the machine's IP by running 'ip addr' or 'ifconfig'\n";
-        break;
-	msg << std::strerror(Error) << "\n" << interfaceInfoToStr(Interface)
-		<< "Bind failure. Check that config file contents for invalid interfaces, and check that there is no other instance of webserv already running with the same config.\n"
-		<< "If bind continues to fail, check elsewhere for occupied interfaces, using ss or netstat & the flags -tulnp";
-      // TODO handle additional errno cases
-      };
-  (void) Info;
-  logging::log(logging::Error, msg.str());
 }
 
 /*
