+++ src/server/ServerInit.cpp
@@ -1,12 +1,10 @@
 #include "Server.hpp"
 #include "Connection.hpp"
 #include "Logging.hpp"
-// #include "NetworkingDefines.hpp" // Can be removed?
 #include <cerrno>  // for errno
 #include <cstdlib> // exit()
 #include <cstring> // for strerror
 #include <exception>
-#include <fcntl.h> // TODO delete before submission, only for debugging
 #include <map>
 #include <ostream>
 #include <poll.h>
@@ -16,10 +14,6 @@
 #include <unistd.h>
 #include <utility>
 #include <vector>
-
-static int clearSocket(const int Sock);
-static int bindToIP(const int Sock, const struct addrinfo *p);
-static int createSocket(const struct addrinfo *P);
 
 
 void Server::initNetworking(const std::list<Website> &Websites) {
@@ -80,7 +74,8 @@ void Server::checkPair(const Listen &Pair){
 }
 
 
-// Check that Port # is uint16_t. Using values outside of this type
+// Check that Port # is uint16_t.
+// This is necessary because using values outside of this type
 // could result in undefined behavior.
 
 void Server::checkPort(const std::string &str){
@@ -142,54 +137,3 @@ int Server::getListeningSocket(struct addrinfo *Info, const Listen &Interface) {
 }
 
 
-static int clearSocket(const int Sock) {
-  int yes = 1;
-  const int ret = setsockopt(Sock, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(int));
-  if (ret == -1) {
-    std::ostringstream msg;
-    msg << "setsockopt: " << std::strerror(errno);
-    logging::log(logging::Error, msg.str());
-  }
-  return (ret);
-}
-
-static int bindToIP(const int Sock, const struct addrinfo *p) {
-
-  const int ret = bind(Sock, p->ai_addr, p->ai_addrlen);
-  if (ret != 0) {
-    close(Sock);
-    std::string msg(gai_strerror(ret));
-    logging::log(logging::Info,
-                 "bind: " + msg + " will continue trying sockets");
-  }
-  return (ret);
-}
-
-static int createSocket(const struct addrinfo *P) {
-  const int sock = socket(P->ai_family, SOCK_STREAM | SOCK_NONBLOCK, 0);
-  // 0 = TCP
-  if (sock == -1) {
-    std::ostringstream msg;
-    msg << "socket: " << std::strerror(errno)
-        << " (will continue trying sockets)";
-    logging::log(logging::Info, msg.str());
-  }
-  //networking::printFcntlFlags(sock);
-  return (sock);
-}
-
-// setToListen() is a wrapper for listen(), which marks the
-// server's socket as a "possive socket". This means it will
-// be used to accept incoming connection requests using accept().
-//
-// The second arguments of listen() is backlog, the maximum
-// length to which the queue of pending connections may grow.
-
-void Server::setToListen(const int Sock) {
-
-  if (listen(Sock, BACKLOG) == -1) {
-    std::ostringstream msg;
-    msg << "listen: " << std::strerror(errno);
-    throw std::runtime_error(msg.str());
-  }
-}
