+++ src/server/ServerWrappers.cpp
@@ -1,0 +1,165 @@
+#include "Logging.hpp"
+#include "Server.hpp"
+#include <cerrno>  // for errno
+#include <cstring> // for strerror
+#include <sys/socket.h>
+#include <sys/types.h>
+#include <unistd.h> // for close
+
+
+/**
+*  \brief getServerInfo() = a wrapper for getaddrinfo()
+* creates an addrinfo struct (standard, from <sys/socket.h>)
+* with information about the website server's own address & our
+* chosen I/O settings. We need this struct to find a usable socket for listening.
+*
+* On error, resulting from library function call failure, an exception is
+* thrown back to main.
+* 
+* \param	Pair	a Listen struct containing the IP:port pair.
+*
+* \return ptr to an allocated addrinfo struct.
+*
+* NOTE: library function getaddrinfo() allocates memory.
+* The addrinfo MUST later be freed by system call freeaddrinfo().
+*
+*/
+
+struct addrinfo *Server::getAddrInfo(const Listen &Pair) {
+
+  struct addrinfo *info;
+  logging::log2(logging::Debug, "getInfo() called with ip: ", Pair.ip);
+  const int ret = getaddrinfo(Pair.ip.c_str(), Pair.port.c_str(), &_hints, &info);
+  if (ret != 0) {
+    const std::string msg(gai_strerror(ret));
+    throw std::runtime_error("getaddrinfo: " + msg);
+  }
+  // logging::log(logging::Debug, addrinfoToStr(info, "server_info:"));
+  return (info);
+}
+
+
+/**
+*  \brief createSocket() = a wrapper for socket() 
+* Creates a socket with the given ddrinfo struct.
+* Logs an error if socket() fails, but does not exit or throw an error,
+* since socket() might fail for some addrinfo structs but succeed for others.
+*
+* The 2nd argument of listen() is BACKLOG, which is the max backlog of pending connections.
+* According to Beej, most sytems silently cap backlog at 20, and 10 is usually sufficient.
+*
+* \param	P	pointer to an addrinfo struct
+* Must have alreadt been generated by getaddrinfo() with the the website server's IP and port.
+*
+* \return fd for new socket, or -1 if socket() fails.
+*/
+
+int Server::createSocket(const struct addrinfo *P) {
+  const int sock = socket(P->ai_family, SOCK_STREAM | SOCK_NONBLOCK, 0);
+  // 0 = TCP
+  if (sock == -1) {
+    std::ostringstream msg;
+    msg << "socket: " << std::strerror(errno)
+        << " (will continue trying sockets)";
+    logging::log(logging::Info, msg.str());
+  }
+  //networking::printFcntlFlags(sock);
+  return (sock);
+}
+
+/**
+*  clearSocket() = a wrapper for setsockopt()
+* This function is meant to clear a socket, just in case the same socket
+* was in use minutes ago, and hasn't yet been systematically cleared by the OS?
+* Logs an error if setsockopt() fails, but does not exit or throw an error,
+* because setsockopt() is not always necessary.
+*
+* NOTE: 
+* I have yet to encounter a situation where it's clearly
+* necessary to clear a socket, but apparently it happens.
+*
+* \param	Sock	fd of socket to be cleared.
+* Must be a socket which has already been created via socket(),
+* but not yet bound to an IP:Port pair.
+*
+* \return: 0 on success, -1 on failure
+*/
+// TODO read more about setsockopt()
+
+int Server::clearSocket(const int Sock) {
+  int yes = 1;
+  const int ret = setsockopt(Sock, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(int));
+  if (ret == -1) {
+    std::ostringstream msg;
+    msg << "setsockopt: " << std::strerror(errno);
+    logging::log(logging::Error, msg.str());
+  }
+  return (ret);
+}
+
+int Server::bindToIP(const int Sock, const struct addrinfo *p) {
+
+  const int ret = bind(Sock, p->ai_addr, p->ai_addrlen);
+  if (ret != 0) {
+    close(Sock);
+    std::string msg(gai_strerror(ret));
+    logging::log(logging::Info,
+                 "bind: " + msg + " will continue trying sockets");
+  }
+  return (ret);
+}
+
+/**
+*  \brief setToListen() = a wrapper for listen() 
+* Marks the server's socket as a "possive socket" (listening socket),
+* which can then be used to accept incoming connections requests via accept().
+* Throws runtime error if listen() fails, with strerror message.
+*
+* The 2nd argument of listen() is BACKLOG, which is the max backlog of pending connections.
+* According to Beej, most sytems silently cap backlog at 20, and 10 is usually sufficient.
+*
+* \param	Sock	fd of socket to be set to listen.
+* Sock must be a socket which has already been created and bound to an IP:Port pair.
+*/
+
+void Server::setToListen(const int Sock) {
+
+  if (listen(Sock, BACKLOG) == -1) {
+    std::ostringstream msg;
+    msg << "listen: " << std::strerror(errno);
+    throw std::runtime_error(msg.str());
+  }
+}
+
+/**
+ * \brief acceptConnection() = a wrapper for accept()
+ * Accepts a connection on a listening socket, and creates a new socket for the incoming connection
+ * Throws runtime error if accept() fails, with strerror message.
+ * 
+ * \param	ListenerFd	fd of the listening socket on which to accept the connection. 
+ * Must be a socket which has already been created, bound to an IP:Port pair, and set to listen.
+ * 
+ * \param	Candidate	pointer to a ClientAddr struct
+ * Will be filled with the new socket fd and client address info for the incoming connection.
+ * 
+ * \return fd for new socket, or -1 if accept() fails.
+ */
+
+int Server::acceptConnection(int ListenerFd, ClientAddr *Candidate) {
+
+  logging::log2(logging::Debug, "acceptConnection() on socket ", ListenerFd);
+  Candidate->clientSock = accept(
+      ListenerFd, (struct sockaddr *)&Candidate->addr, &Candidate->addrSize);
+  if (Candidate->clientSock == -1) {
+    std::ostringstream msg;
+    msg << "accept: " << std::strerror(errno)
+        << " (can continue trying to accept connections)";
+    logging::log(logging::Error,
+                 msg.str());
+    return (-1);
+  }
+  logging::log2(logging::Debug, "Connection accepted on socket ",
+                Candidate->clientSock);
+  // printFcntlFlags(Candidate->clientSock); // demonstrates that individual client sockets are blocking
+  return (0);
+}
